Moved functions to helpers file. Minor cleanup.
